+++ blob/client/client_test.go
@@ -99,7 +99,7 @@ var _ = Describe("Client", func() {
 			}
 		})
 
-		clientAssertions := func() {
+		authorizeAssertions := func() {
 			Context("with user id", func() {
 				var userID string
 
@@ -111,7 +111,7 @@ var _ = Describe("Client", func() {
 					var filter *blob.Filter
 					var pagination *page.Pagination
 
-					listAssertions := func() {
+					parameterAssertions := func() {
 						Context("without server response", func() {
 							AfterEach(func() {
 								Expect(server.ReceivedRequests()).To(BeEmpty())
@@ -239,7 +239,7 @@ var _ = Describe("Client", func() {
 							requestHandlers = append(requestHandlers, VerifyRequest(http.MethodGet, fmt.Sprintf("/v1/users/%s/blobs", userID), ""))
 						})
 
-						listAssertions()
+						parameterAssertions()
 					})
 
 					When("the request has random filter and pagination parameters", func() {
@@ -252,10 +252,10 @@ var _ = Describe("Client", func() {
 								"page":      []string{strconv.Itoa(pagination.Page)},
 								"size":      []string{strconv.Itoa(pagination.Size)},
 							}
-							requestHandlers = append(requestHandlers, VerifyRequest("GET", fmt.Sprintf("/v1/users/%s/blobs", userID), query.Encode()))
+							requestHandlers = append(requestHandlers, VerifyRequest(http.MethodGet, fmt.Sprintf("/v1/users/%s/blobs", userID), query.Encode()))
 						})
 
-						listAssertions()
+						parameterAssertions()
 					})
 				})
 
@@ -310,11 +310,11 @@ var _ = Describe("Client", func() {
 						})
 					})
 
-					createAssertions := func() {
+					digestAssertions := func() {
 						Context("with server response", func() {
 							BeforeEach(func() {
 								requestHandlers = append(requestHandlers,
-									VerifyRequest("POST", fmt.Sprintf("/v1/users/%s/blobs", userID)),
+									VerifyRequest(http.MethodPost, fmt.Sprintf("/v1/users/%s/blobs", userID)),
 									VerifyContentType(*content.MediaType),
 									VerifyBody(body),
 								)
@@ -382,7 +382,7 @@ var _ = Describe("Client", func() {
 							content.DigestMD5 = nil
 						})
 
-						createAssertions()
+						digestAssertions()
 					})
 
 					When("the request has a digest header", func() {
@@ -390,7 +390,7 @@ var _ = Describe("Client", func() {
 							requestHandlers = append(requestHandlers, VerifyHeaderKV("Digest", fmt.Sprintf("MD5=%s", *content.DigestMD5)))
 						})
 
-						createAssertions()
+						digestAssertions()
 					})
 				})
 			})
@@ -433,7 +433,7 @@ var _ = Describe("Client", func() {
 					Context("with server response", func() {
 						BeforeEach(func() {
 							requestHandlers = append(requestHandlers,
-								VerifyRequest("GET", fmt.Sprintf("/v1/blobs/%s", id)),
+								VerifyRequest(http.MethodGet, fmt.Sprintf("/v1/blobs/%s", id)),
 								VerifyContentType(""),
 								VerifyBody(nil),
 							)
@@ -504,30 +504,30 @@ var _ = Describe("Client", func() {
 
 						It("returns an error when the context is missing", func() {
 							ctx = nil
-							content, err := client.GetContent(ctx, id)
+							result, err := client.GetContent(ctx, id)
 							errorsTest.ExpectEqual(err, errors.New("context is missing"))
-							Expect(content).To(BeNil())
+							Expect(result).To(BeNil())
 						})
 
 						It("returns an error when the id is missing", func() {
 							id = ""
-							content, err := client.GetContent(ctx, id)
+							result, err := client.GetContent(ctx, id)
 							errorsTest.ExpectEqual(err, errors.New("id is missing"))
-							Expect(content).To(BeNil())
+							Expect(result).To(BeNil())
 						})
 
 						It("returns an error when the id is invalid", func() {
 							id = "invalid"
-							content, err := client.GetContent(ctx, id)
+							result, err := client.GetContent(ctx, id)
 							errorsTest.ExpectEqual(err, errors.New("id is invalid"))
-							Expect(content).To(BeNil())
+							Expect(result).To(BeNil())
 						})
 					})
 
 					Context("with server response", func() {
 						BeforeEach(func() {
 							requestHandlers = append(requestHandlers,
-								VerifyRequest("GET", fmt.Sprintf("/v1/blobs/%s/content", id)),
+								VerifyRequest(http.MethodGet, fmt.Sprintf("/v1/blobs/%s/content", id)),
 								VerifyContentType(""),
 								VerifyBody(nil),
 							)
@@ -543,9 +543,9 @@ var _ = Describe("Client", func() {
 							})
 
 							It("returns an error", func() {
-								content, err := client.GetContent(ctx, id)
+								result, err := client.GetContent(ctx, id)
 								errorsTest.ExpectEqual(err, request.ErrorUnauthenticated())
-								Expect(content).To(BeNil())
+								Expect(result).To(BeNil())
 							})
 						})
 
@@ -555,9 +555,9 @@ var _ = Describe("Client", func() {
 							})
 
 							It("returns an error", func() {
-								content, err := client.GetContent(ctx, id)
+								result, err := client.GetContent(ctx, id)
 								errorsTest.ExpectEqual(err, request.ErrorUnauthorized())
-								Expect(content).To(BeNil())
+								Expect(result).To(BeNil())
 							})
 						})
 
@@ -566,10 +566,10 @@ var _ = Describe("Client", func() {
 								requestHandlers = append(requestHandlers, RespondWithJSONEncoded(http.StatusNotFound, errors.NewSerializable(request.ErrorResourceNotFoundWithID(id)), responseHeaders))
 							})
 
-							It("returns an error", func() {
-								content, err := client.GetContent(ctx, id)
+							It("returns successfully without result", func() {
+								result, err := client.GetContent(ctx, id)
 								Expect(err).ToNot(HaveOccurred())
-								Expect(content).To(BeNil())
+								Expect(result).To(BeNil())
 							})
 						})
 
@@ -582,9 +582,9 @@ var _ = Describe("Client", func() {
 							})
 
 							It("returns successfully", func() {
-								content, err := client.GetContent(ctx, id)
+								result, err := client.GetContent(ctx, id)
 								errorsTest.ExpectEqual(err, request.ErrorHeaderInvalid("Digest"))
-								Expect(content).To(BeNil())
+								Expect(result).To(BeNil())
 							})
 						})
 
@@ -601,9 +601,9 @@ var _ = Describe("Client", func() {
 							})
 
 							It("returns successfully", func() {
-								content, err := client.GetContent(ctx, id)
+								result, err := client.GetContent(ctx, id)
 								errorsTest.ExpectEqual(err, request.ErrorHeaderInvalid("Content-Type"))
-								Expect(content).To(BeNil())
+								Expect(result).To(BeNil())
 							})
 						})
 
@@ -624,12 +624,13 @@ var _ = Describe("Client", func() {
 							})
 
 							It("returns successfully", func() {
-								content, err := client.GetContent(ctx, id)
+								result, err := client.GetContent(ctx, id)
 								Expect(err).ToNot(HaveOccurred())
-								Expect(content).ToNot(BeNil())
-								Expect(content.Body).ToNot(BeNil())
-								Expect(content.DigestMD5).To(Equal(&digestMD5))
-								Expect(content.MediaType).To(Equal(&mediaType))
+								Expect(result).ToNot(BeNil())
+								Expect(result.Body).ToNot(BeNil())
+								defer result.Body.Close()
+								Expect(*result.DigestMD5).To(Equal(digestMD5))
+								Expect(*result.MediaType).To(Equal(mediaType))
 							})
 						})
 					})
@@ -676,7 +677,7 @@ var _ = Describe("Client", func() {
 						})
 					})
 
-					deleteAssertions := func() {
+					conditionAssertions := func() {
 						Context("with server response", func() {
 							AfterEach(func() {
 								Expect(server.ReceivedRequests()).To(HaveLen(1))
@@ -711,7 +712,7 @@ var _ = Describe("Client", func() {
 									requestHandlers = append(requestHandlers, RespondWithJSONEncoded(http.StatusNotFound, errors.NewSerializable(request.ErrorResourceNotFoundWithID(id)), responseHeaders))
 								})
 
-								It("returns successfully with delete false", func() {
+								It("returns successfully with deleted false", func() {
 									deleted, err := client.Delete(ctx, id, condition)
 									Expect(err).ToNot(HaveOccurred())
 									Expect(deleted).To(BeFalse())
@@ -736,26 +737,26 @@ var _ = Describe("Client", func() {
 						BeforeEach(func() {
 							condition = nil
 							requestHandlers = append(requestHandlers,
-								VerifyRequest("DELETE", fmt.Sprintf("/v1/blobs/%s", id)),
+								VerifyRequest(http.MethodDelete, fmt.Sprintf("/v1/blobs/%s", id)),
 								VerifyContentType(""),
 								VerifyBody(nil),
 							)
 						})
 
-						deleteAssertions()
+						conditionAssertions()
 					})
 
 					When("condition revision is missing", func() {
 						BeforeEach(func() {
 							condition.Revision = nil
 							requestHandlers = append(requestHandlers,
-								VerifyRequest("DELETE", fmt.Sprintf("/v1/blobs/%s", id)),
+								VerifyRequest(http.MethodDelete, fmt.Sprintf("/v1/blobs/%s", id)),
 								VerifyContentType(""),
 								VerifyBody(nil),
 							)
 						})
 
-						deleteAssertions()
+						conditionAssertions()
 					})
 
 					When("condition revision is present", func() {
@@ -764,13 +765,13 @@ var _ = Describe("Client", func() {
 								"revision": []string{strconv.Itoa(*condition.Revision)},
 							}
 							requestHandlers = append(requestHandlers,
-								VerifyRequest("DELETE", fmt.Sprintf("/v1/blobs/%s", id), query.Encode()),
+								VerifyRequest(http.MethodDelete, fmt.Sprintf("/v1/blobs/%s", id), query.Encode()),
 								VerifyContentType(""),
 								VerifyBody(nil),
 							)
 						})
 
-						deleteAssertions()
+						conditionAssertions()
 					})
 				})
 			})
@@ -783,7 +784,7 @@ var _ = Describe("Client", func() {
 				requestHandlers = append(requestHandlers, VerifyHeaderKV("X-Tidepool-Service-Secret", config.ServiceSecret))
 			})
 
-			clientAssertions()
+			authorizeAssertions()
 		})
 
 		When("client must authorize as user", func() {
@@ -794,7 +795,7 @@ var _ = Describe("Client", func() {
 				ctx = request.NewContextWithDetails(ctx, request.NewDetails(request.MethodAccessToken, userTest.RandomID(), sessionToken))
 			})
 
-			clientAssertions()
+			authorizeAssertions()
 		})
 	})
 })