+++ data/source/source.go
@@ -32,16 +32,14 @@ func States() []string {
 	}
 }
 
-type Accessor interface {
+// FUTURE: Add DeleteAll
+
+type Client interface {
 	List(ctx context.Context, userID string, filter *Filter, pagination *page.Pagination) (Sources, error)
 	Create(ctx context.Context, userID string, create *Create) (*Source, error)
 	Get(ctx context.Context, id string) (*Source, error)
 	Update(ctx context.Context, id string, condition *request.Condition, update *Update) (*Source, error)
 	Delete(ctx context.Context, id string, condition *request.Condition) (bool, error)
-}
-
-type Client interface {
-	Accessor
 }
 
 type Filter struct {