+++ data/source/store/structured/structured.go
@@ -1,9 +1,12 @@
 package structured
 
 import (
+	"context"
 	"io"
 
 	dataSource "github.com/tidepool-org/platform/data/source"
+	"github.com/tidepool-org/platform/page"
+	"github.com/tidepool-org/platform/request"
 )
 
 type Store interface {
@@ -12,5 +15,10 @@ type Store interface {
 
 type Session interface {
 	io.Closer
-	dataSource.Accessor
+
+	List(ctx context.Context, userID string, filter *dataSource.Filter, pagination *page.Pagination) (dataSource.Sources, error)
+	Create(ctx context.Context, userID string, create *dataSource.Create) (*dataSource.Source, error)
+	Get(ctx context.Context, id string) (*dataSource.Source, error)
+	Update(ctx context.Context, id string, condition *request.Condition, update *dataSource.Update) (*dataSource.Source, error)
+	Destroy(ctx context.Context, id string, condition *request.Condition) (bool, error)
 }